Break out minimap
